refactor: 회원가입 링크 연결 및 로그아웃 알림창 수정
