+++ src/components/Sidebar/index.js
@@ -1,11 +1,15 @@
-import { Wrapper } from "./style";
+import { useDispatch } from "react-redux";
+import { useNavigate } from "react-router-dom";
+
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
-import { useDispatch } from "react-redux";
+
+import { Wrapper } from "./style";
 import { setModalOpen } from "../../store/slices/modalSlice";
 
 function Sidebar({ setIsOpen, role }) {
   const dispatch = useDispatch();
+  const navigate = useNavigate();
 
   return (
     <Wrapper>
@@ -14,7 +18,11 @@ function Sidebar({ setIsOpen, role }) {
         className="arrow-left"
         onClick={() => setIsOpen(false)}
       />
-      {role === "GUEST" && <div className="content1">회원가입</div>}
+      {role === "GUEST" && (
+        <div className="content1" onClick={() => navigate("/signup")}>
+          회원가입
+        </div>
+      )}
       {role === "MEMBER" && (
         <div>
           <div
@@ -30,15 +38,6 @@ function Sidebar({ setIsOpen, role }) {
       )}
       {role === "ADMIN" && (
         <div>
-          <div className="content1">그룹 관리하기</div>
-          <div
-            className="content2"
-            onClick={() =>
-              dispatch(setModalOpen({ type: "createNotice", message: "" }))
-            }
-          >
-            그룹 공지 보내기
-          </div>
           <div
             className="content1"
             onClick={() =>
@@ -47,7 +46,14 @@ function Sidebar({ setIsOpen, role }) {
           >
             그룹 관리하기
           </div>
-          <div className="content2">그룹 공지 보내기</div>
+          <div
+            className="content2"
+            onClick={() =>
+              dispatch(setModalOpen({ type: "createNotice", message: "" }))
+            }
+          >
+            그룹 공지 보내기
+          </div>
         </div>
       )}
     </Wrapper>
